split `BottomPanelSliders` in three
