+++ src/aics-image-viewer/components/dimension_sliders/styles.css
